crashpad_pid_check.patch
When both browser process and renderer process are connecting to the pipe,
the API may return the PID of browser process as real_pid, which is different
from the PID of renderer process.

This is caused by the crashReporter getting started after the sanbox, after
we redesign crashReporter's API to make it alwasy start before the
sanbox, we can remove this patch.

See following links for more:
https://github.com/electron/electron/pull/18483#discussion_r292703588
https://github.com/electron/electron/pull/18483#issuecomment-501090683

Patch-Dir: src/electron/patches/chromium
Patch-Filename: crashpad_pid_check.patch