feat: add support for overriding the base spellchecker download URL
This patch is required as the testing-only method we were using does not
take into account the dictionary name and therefore will not work for
production use cases.  This is unlikely to be upstreamed as the change
is entirely in //chrome.

Patch-Dir: src/electron/patches/chromium
Patch-Filename: feat_add_support_for_overriding_the_base_spellchecker_download_url.patch