feat: filter out non-shareable windows in the current application in ScreenCaptureKitDevice
This patch ensures that windows protected via win.setContentProtection(true) do not appear in full display captures via desktopCapturer.  This patch could be upstreamed but as the check is limited to in-process windows it doesn't make a lot of sense for Chromium itself.  This patch currently has a limitation that it only function for windows created / protected BEFORE the stream is started.  There is theoretical future work we can do via polling / observers to automatically update the SCContentFilter when new windows are made but for now this will solve 99+% of the problem and folks can re-order their logic a bit to get it working for their use cases.

Patch-Dir: src/electron/patches/chromium
Patch-Filename: feat_filter_out_non-shareable_windows_in_the_current_application_in.patch