disable freezing flags after init in node
This was introduced in
https://chromium-review.googlesource.com/c/chromium/src/+/3687671.

When running node in the renderer, flags are updated after initialization, so
freezing the flags in Blink causes node initialization to fail.

If possible, it would be ideal to do this without a patch.
https://bugs.chromium.org/p/v8/issues/detail?id=12887 suggests that there may
at some point be an API to "unfreeze" the flags, or we may be able to refactor
node initialization to not update flags after V8 initialization.

Patch-Dir: src/electron/patches/chromium
Patch-Filename: disable_freezing_flags_after_init_in_node.patch