fix: activate background material on windows
This patch adds a condition to the HWND message handler to allow windows
with translucent background materials to become activated.

It also ensures the lParam of WM_NCACTIVATE is set to -1 so as to not repaint
the client area, which can lead to a title bar incorrectly being displayed in
frameless windows.

This patch likely can't be upstreamed as-is, as Chromium doesn't have
this use case in mind currently.

Patch-Dir: src/electron/patches/chromium
Patch-Filename: fix_activate_background_material_on_windows.patch