feat: enable customizing symbol color in FrameCaptionButton
This enables customizing the symbol color on a given FrameCaptionButton
for the Window Controls Overlay API on Linux. By default, the symbol color
is dynamically calculated based on the background color of the button to
ensure it has minimum contrast required to be accessible.

This should be upstreamed to Chromium if possible.

Patch-Dir: src/electron/patches/chromium
Patch-Filename: feat_enable_customizing_symbol_color_in_framecaptionbutton.patch