chore: partial revert of https://chromium-review.googlesource.com/c/chromium/src/+/5894233
The `CHECK_EQ()` being removed here is failing in the Chromium roll.
I experimented with backporting the `CHECK_EQ()` to `main` and it's
failing there as well, so some pre-existing behavior in Electron is
incompatible with this assertion.

I will file an upgrades-followup-task issue in Electron and try to
track down the source of this problem & figure out if we can fix it
by changing something in Electron.

Patch-Dir: src/electron/patches/chromium
Patch-Filename: chore_partial_revert_of.patch