Support BSTR PKEY_AppUserModel_ID in Windows Shortcuts
It's possible for a shortcut link on Windows to be BSTR format;
this resulted in a DCHECK previously but as of the NOTREACHED
fatal migration now will crash a given app. This has created an
issue for Electron, which allows providing the shortcut information
to end users. This CL thus allows for supporting VT_BSTR variants
for PKEY_AppUserModel_ID and mitigates a crash. One such example
is https://github.com/wez/wezterm, which after this change
correctly gets an appUserModelId of 'org.wezfurlong.wezterm'.

Bug: N/A
Change-Id: I5ad114bc345059637465b68a53a2f3f8c42de898

Patch-Dir: src/electron/patches/chromium
Patch-Filename: support_bstr_pkey_appusermodel_id_in_windows_shortcuts.patch